+++ protocols/kad/src/behaviour.rs
@@ -23,7 +23,6 @@
 mod test;
 
 use crate::addresses::Addresses;
-use crate::bootstrap::Status;
 use crate::handler::{Handler, HandlerEvent, HandlerIn, RequestId};
 use crate::jobs::*;
 use crate::kbucket::{self, Distance, KBucketsTable, NodeStatus};
@@ -185,7 +184,8 @@ pub struct Config {
     provider_publication_interval: Option<Duration>,
     kbucket_inserts: BucketInserts,
     caching: Caching,
-    bootstrap_interval: Option<Duration>,
+    periodic_bootstrap_interval: Option<Duration>,
+    automatic_bootstrap_throttle: Option<Duration>,
 }
 
 impl Default for Config {
@@ -202,7 +202,8 @@ impl Default for Config {
             provider_record_ttl: Some(Duration::from_secs(24 * 60 * 60)),
             kbucket_inserts: BucketInserts::OnConnected,
             caching: Caching::Enabled { max_peers: 1 },
-            bootstrap_interval: Some(Duration::from_secs(5 * 60)),
+            periodic_bootstrap_interval: Some(Duration::from_secs(5 * 60)),
+            automatic_bootstrap_throttle: Some(Duration::from_millis(10)),
         }
     }
 }
@@ -398,11 +399,27 @@ impl Config {
         self
     }
 
-    /// Sets the interval on which [`Behaviour::bootstrap`] is called from [`Behaviour::poll`]
+    /// Sets the interval on which [`Behaviour::bootstrap`] is called periodically.
     ///
-    /// `None` means we don't bootstrap at all.
-    pub fn set_bootstrap_interval(&mut self, interval: Option<Duration>) -> &mut Self {
-        self.bootstrap_interval = interval;
+    /// * Default to `5` minutes.
+    /// * Set to `None` to disable periodic bootstrap.
+    pub fn set_periodic_bootstrap_interval(&mut self, interval: Option<Duration>) -> &mut Self {
+        self.periodic_bootstrap_interval = interval;
+        self
+    }
+
+    /// Sets the time to wait before calling [`Behaviour::bootstrap`] after a new peer is inserted in the routing table.
+    /// This prevent cascading bootstrap requests when multiple peers are inserted into the routing table "at the same time".
+    /// This also allows to wait a little bit for other potential peers to be inserted into the routing table before
+    /// triggering a bootstrap, giving more context to the future bootstrap request.  
+    ///
+    /// * Default to `10` ms.
+    /// * Set to `Some(Duration::ZERO)` to never wait before triggering a bootstrap request when a new peer
+    ///     is inserted in the routing table.
+    /// * Set to `None` to disable automatic bootstrap (no bootstrap request will be triggered when a new
+    ///     peer is inserted in the routing table).
+    pub fn set_automatic_bootstrap_throttle(&mut self, duration: Option<Duration>) -> &mut Self {
+        self.automatic_bootstrap_throttle = duration;
         self
     }
 }
@@ -462,7 +479,10 @@ where
             mode: Mode::Client,
             auto_mode: true,
             no_events_waker: None,
-            bootstrap_status: Status::new(config.bootstrap_interval),
+            bootstrap_status: bootstrap::Status::new(
+                config.periodic_bootstrap_interval,
+                config.automatic_bootstrap_throttle,
+            ),
         }
     }
 
@@ -883,11 +903,14 @@ where
     ///
     /// > **Note**: Bootstrapping requires at least one node of the DHT to be known.
     /// > See [`Behaviour::add_address`].
-    /// > **Note**: Bootstrap does not require to be called manually. It is automatically
-    /// invoked at regular intervals based on the configured bootstrapping interval.
-    /// The bootstrapping interval is used to call bootstrap periodically
+    ///
+    /// > **Note**: Bootstrap does not require to be called manually. It is periodically
+    /// invoked at regular intervals based on the configured `periodic_bootstrap_interval` (see
+    /// [`Config::set_periodic_bootstrap_interval`] for details) and it is also automatically invoked
+    /// when a new peer is inserted in the routing table based on the `automatic_bootstrap_throttle`
+    /// (see [`Config::set_automatic_bootstrap_throttle`] for details).
+    /// These two config parameters are used to call [`Behaviour::bootstrap`] periodically and automatically
     /// to ensure a healthy routing table.
-    /// > See [`Config::set_bootstrap_interval`] for details.
     pub fn bootstrap(&mut self) -> Result<QueryId, NoKnownPeers> {
         let local_key = self.kbuckets.local_key().clone();
         let info = QueryInfo::Bootstrap {
@@ -897,15 +920,12 @@ where
         };
         let peers = self.kbuckets.closest_keys(&local_key).collect::<Vec<_>>();
         if peers.is_empty() {
-            return Err(NoKnownPeers());
+            Err(NoKnownPeers())
+        } else {
+            self.bootstrap_status.on_started();
+            let inner = QueryInner::new(info);
+            Ok(self.queries.add_iter_closest(local_key, peers, inner))
         }
-
-        let inner = QueryInner::new(info);
-        let id = self.queries.add_iter_closest(local_key, peers, inner);
-
-        self.bootstrap_status.on_started();
-
-        Ok(id)
     }
 
     /// Establishes the local node as a provider of a value for the given key.
@@ -1415,7 +1435,7 @@ where
                         .continue_iter_closest(query_id, target.clone(), peers, inner);
                 } else {
                     step.last = true;
-                    self.bootstrap_status.on_result(&result.stats);
+                    self.bootstrap_status.on_finish();
                 };
 
                 Some(Event::OutboundQueryProgressed {
@@ -1618,7 +1638,7 @@ where
                         .continue_iter_closest(query_id, target.clone(), peers, inner);
                 } else {
                     step.last = true;
-                    self.bootstrap_status.on_result(&result.stats);
+                    self.bootstrap_status.on_finish();
                 }
 
                 Some(Event::OutboundQueryProgressed {
@@ -2491,7 +2511,7 @@ where
             self.put_record_job = Some(job);
         }
 
-        // Poll bootstrap periodically.
+        // Poll bootstrap periodically and automatically.
         if let Poll::Ready(()) = self.bootstrap_status.poll_next_bootstrap(cx) {
             if let Err(e) = self.bootstrap() {
                 tracing::warn!("Failed to trigger bootstrap: {e}");